+++ R/fosr_bayes.R
@@ -72,7 +72,6 @@
 #' @import mgcv
 #' @import splines2
 #' @import rstan
-#' @importFrom brms brmsformula
 #' @importFrom refund fpca.face
 #' @importFrom stats gaussian
 #' @export fosr_bayes